Try to make a good habit to avoid writes to the wrong place.
This commit makes Profile attributes that are hard to get right
protected, and adds two methods to convert relative paths to paths
suitable for reading (src_path(), from the parent's top_dir) and for
writing (dst_path(), to the profile's own top_dir). It also renames
variables to have src_ or dst_ prefixes, so that it should be harder
to accidentally use a path in the wrong way.