Give callbacks access to the profile that called them.
I originally thought to avoid that so that it was harder for a
callback to write to the parent. But now that src_path and dst_path
exist, that should be easier to avoid, and giving the profile should
help with logging, and with select callbacks getting access to file
contents.